pyproject.toml
[tool.poetry]
name = "binance-notificator"
version = "0.1.0"
description = ""
authors = ["Emre MENTEŞE <emrementese@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
mashumaro = "^3.10"
colorama = "^0.4.6"
binance-connector = "^3.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
